Read from stream buffer if available
On python 3 we're still facing random stream processing errors. A recent
patch on stackviz (a different project that consumes subunit) was trying
to fix a similar issue by using the stream's buffer if available instead
of the stream itself. This commit attempts the same thing in an attempt
to resolve the issue.